backend/requirements_backend.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
loguru>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.8.2

langchain>=0.3.7
langchain-core>=0.3.15
langchain-openai>=0.2.3
openai>=1.56.0
tiktoken>=0.5.2
chromadb>=0.4.22
faiss-cpu>=1.12.0

nltk>=3.9.1
spacy>=3.7.2

requests>=2.32.3
beautifulsoup4>=4.12.3
httpx>=0.27.0

google-generativeai
psycopg2-binary

sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic-settings>=2.3.0
python-multipart>=0.0.6
tqdm>=4.66.4
sentence-transformers>=2.7.0

pytest
pytest-asyncio
httpx
faker
locust
pytest-benchmark

